Move RunsTable into CallsTable file
